Move airline association to routes
