+++ client/src/components/admin/FlightManagement.js
@@ -26,7 +26,6 @@ import {
 } from '../../redux/actions/flight';
 import { fetchTariffs, deleteTariff } from '../../redux/actions/tariff';
 import { fetchRoutes } from '../../redux/actions/route';
-import { fetchAirlines } from '../../redux/actions/airline';
 import { FIELD_TYPES, createAdminManager, formatDateTime } from './utils';
 import {
 	ENUM_LABELS,
@@ -44,10 +43,7 @@ const FlightManagement = () => {
 	const { routes, isLoading: routesLoading } = useSelector(
 		(state) => state.routes
 	);
-	const { airlines, isLoading: airlinesLoading } = useSelector(
-		(state) => state.airlines
-	);
-	const { tariffs } = useSelector((state) => state.tariffs);
+        const { tariffs } = useSelector((state) => state.tariffs);
 
 	const [tariffDialogOpen, setTariffDialogOpen] = useState(false);
 	const [tariffAction, setTariffAction] = useState(null);
@@ -56,9 +52,8 @@ const FlightManagement = () => {
 
 	useEffect(() => {
 		dispatch(fetchFlights());
-		dispatch(fetchRoutes());
-		dispatch(fetchAirlines());
-		dispatch(fetchTariffs());
+                dispatch(fetchRoutes());
+                dispatch(fetchTariffs());
 	}, [dispatch]);
 
 	const getRouteOptions = () => {
@@ -78,25 +73,7 @@ const FlightManagement = () => {
 		return routes.find((route) => route.id === id);
 	};
 
-	const getAirlineOptions = () => {
-		if (!airlines || !Array.isArray(airlines)) {
-			return [];
-		}
-		return airlines.map((airline) => ({
-			value: airline.id,
-			label: `${airline.name} (${airline.iata_code})`,
-		}));
-	};
-
-	const getAirlineById = (id) => {
-		if (!airlines || !Array.isArray(airlines)) {
-			return null;
-		}
-		return airlines.find((airline) => airline.id === id);
-	};
-
-	const routeOptions = useMemo(() => getRouteOptions(), [routes]);
-	const airlineOptions = useMemo(() => getAirlineOptions(), [airlines]);
+        const routeOptions = useMemo(() => getRouteOptions(), [routes]);
 
 	const [deleteTariffDialog, setDeleteTariffDialog] = useState({
 		open: false,
@@ -152,19 +129,6 @@ const FlightManagement = () => {
 			},
 			validate: (value) =>
 				!value ? VALIDATION_MESSAGES.FLIGHT.route_id.REQUIRED : null,
-		},
-		airlineId: {
-			key: 'airlineId',
-			apiKey: 'airline_id',
-			label: FIELD_LABELS.FLIGHT.airline_id,
-			type: FIELD_TYPES.SELECT,
-			options: airlineOptions,
-			formatter: (value) => {
-				const airline = getAirlineById(value);
-				return airline ? `${airline.iata_code}` : value;
-			},
-			validate: (value) =>
-				!value ? VALIDATION_MESSAGES.FLIGHT.airline_id.REQUIRED : null,
 		},
 		status: {
 			key: 'status',
@@ -382,7 +346,7 @@ const FlightManagement = () => {
 				onDelete={handleDeleteFlight}
 				renderForm={adminManager.renderForm}
 				addButtonText={UI_LABELS.ADMIN.modules.flights.add_button}
-				isLoading={isLoading || routesLoading || airlinesLoading}
+                                isLoading={isLoading || routesLoading}
 				error={errors}
 			/>
