+++ client/src/constants/fieldLabels.js
@@ -24,17 +24,16 @@ export const FIELD_LABELS = {
         city_code: 'Код города',
         country_id: 'Страна',
     },
-	FLIGHT: {
-		flight_number: 'Номер рейса',
-		route_id: 'Маршрут рейса',
-		airline_id: 'Авиакомпания',
-		origin_airport_id: 'Аэропорт отправления',
-		destination_airport_id: 'Аэропорт прибытия',
-		scheduled_departure: 'Время отправления',
-		scheduled_arrival: 'Время прибытия',
-		status: 'Статус рейса',
-		tariffs: 'Тарифы',
-	},
+        FLIGHT: {
+                flight_number: 'Номер рейса',
+                route_id: 'Маршрут рейса',
+                origin_airport_id: 'Аэропорт отправления',
+                destination_airport_id: 'Аэропорт прибытия',
+                scheduled_departure: 'Время отправления',
+                scheduled_arrival: 'Время прибытия',
+                status: 'Статус рейса',
+                tariffs: 'Тарифы',
+        },
 	PASSENGER: {
 		first_name: 'Имя',
 		last_name: 'Фамилия',
@@ -74,13 +73,14 @@ export const FIELD_LABELS = {
 		code_a2: 'Код A2',
 		code_a3: 'Код A3',
 	},
-	ROUTE: {
-		flight_number: 'Номер рейса',
-		origin_airport_id: 'Аэропорт отправления',
-		destination_airport_id: 'Аэропорт прибытия',
-		departure_time: 'Время отправления',
-		arrival_time: 'Время прибытия',
-	},
+        ROUTE: {
+                flight_number: 'Номер рейса',
+                airline_id: 'Авиакомпания',
+                origin_airport_id: 'Аэропорт отправления',
+                destination_airport_id: 'Аэропорт прибытия',
+                departure_time: 'Время отправления',
+                arrival_time: 'Время прибытия',
+        },
 	TARIFF: {
 		name: 'Название тарифа',
 		description: 'Описание',
